to_os_path() moved from os_tools to misc
